Move test to existing test file
